Added cat command
The cat command prints repository objects to stdout. All objects
can be inspected except for the key files, which are encrypted with
the password, not with the repository master key.
