+++ src/commands/mod.rs
@@ -14,7 +14,52 @@
 // You should have received a copy of the GNU General Public License
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
+use clap::{Parser, Subcommand};
+
+pub mod cat;
 pub mod commit;
 pub mod init;
 pub mod log;
 pub mod restore;
+
+// CLI arguments
+#[derive(Parser, Debug)]
+#[clap(
+    version = env!("CARGO_PKG_VERSION"), // Version from crate metadata
+    about = "Incremental backup tool"
+)]
+pub struct Cli {
+    // Subcommand
+    #[command(subcommand)]
+    pub command: Command,
+
+    // Global arguments
+    #[clap(flatten)]
+    pub global_args: GlobalArgs,
+}
+
+// List of commands
+#[derive(Subcommand, Debug)]
+pub enum Command {
+    #[clap(about = "Initialize a new repository")]
+    Init(init::CmdArgs),
+
+    #[clap(about = "Show all snapshots present in the repository")]
+    Log(log::CmdArgs),
+
+    #[clap(about = "Create a new snapshot")]
+    Commit(commit::CmdArgs),
+
+    #[clap(about = "Restores a snapshot")]
+    Restore(restore::CmdArgs),
+
+    #[clap(about = "Prints repository objects")]
+    Cat(cat::CmdArgs),
+}
+
+#[derive(Parser, Debug)]
+pub struct GlobalArgs {
+    /// Repository path
+    #[clap(short, long, value_parser)]
+    pub repo: String,
+}
