+++ src/main.rs
@@ -29,17 +29,18 @@ use anyhow::Result;
 use clap::Parser;
 use colored::Colorize;
 
-fn run(args: &cli::Cli) -> Result<()> {
+fn run(args: &commands::Cli) -> Result<()> {
     match &args.command {
-        cli::Command::Init(cmd_args) => commands::init::run(&args.global_args, cmd_args),
-        cli::Command::Log(cmd_args) => commands::log::run(&args.global_args, cmd_args),
-        cli::Command::Commit(cmd_args) => commands::commit::run(&args.global_args, cmd_args),
-        cli::Command::Restore(cmd_args) => commands::restore::run(&args.global_args, cmd_args),
+        commands::Command::Init(cmd_args) => commands::init::run(&args.global_args, cmd_args),
+        commands::Command::Log(cmd_args) => commands::log::run(&args.global_args, cmd_args),
+        commands::Command::Commit(cmd_args) => commands::commit::run(&args.global_args, cmd_args),
+        commands::Command::Restore(cmd_args) => commands::restore::run(&args.global_args, cmd_args),
+        commands::Command::Cat(cmd_args) => commands::cat::run(&args.global_args, cmd_args),
     }
 }
 
 fn main() -> Result<()> {
-    let args = cli::Cli::parse();
+    let args = commands::Cli::parse();
 
     // Run the command
     if let Err(e) = run(&args) {
