+++ src/repository/tree.rs
@@ -31,8 +31,8 @@ use serde::{Deserialize, Serialize};
 use crate::archiver::Archiver;
 
 use super::{
-    repository::{ObjectId, RepositoryBackend},
     storage::SecureStorage,
+    {ObjectId, RepositoryBackend},
 };
 
 /// Node metadata. This struct is serialized; keep field order stable.
